sphinx: move HopData to path.go re-add its Decode method
In this commit, we move the HopData struct back into path.go. Along the
way, we remove its old HMAC field (since it's now in the hop payload),
and also restore its Decode method for later use.
